Fix piano.

+++ sampler.h
@@ -7,7 +7,7 @@
 struct Sample {
     Map data; //Sample Definition
     int16 trigger=0; int16 lovel=0; int16 hivel=127; int16 lokey=0; int16 hikey=127; //Input Controls
-    int16 pitch_keycenter=60; int32 release=48000; int16 amp_veltrack=100; int16 rt_decay=0; float volume=1; //Performance Parameters
+    int16 pitch_keycenter=60; int32 release=48000; int16 amp_veltrack=100; int16 rt_decay=0; int16 volume=1; //Performance Parameters
 };
 
 struct Note : FLAC { int remaining; int release; int key; int layer; int velocity; float level; };
